CIFS: Return error code when getting file handle for writeback
Now we just return NULL cifsFileInfo pointer in cases we didn't find
or couldn't reopen a file. This hides errors from cifs_reopen_file()
which might be an indication to do a retry instead of failing
an operation. Create new get_writable_file() routine that returns
error codes from cifs_reopen_file() and use it in the writeback
codepath.

Signed-off-by: Pavel Shilovsky <pshilov@microsoft.com>